requirements.txt
torch
diffusers
numpy
matplotlib
transformers
accelerate
xformers
mediapy
fire
decord
OpenEXR
Pillow
pandas
pytorch3d
kornia
opencv-python
h5py
moderngl
piqp
gin-config
imageio
wandb
wandb-osh
scipy
imageio-ffmpeg
# imageio-pyav